requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asteval==1.0.6
astunparse==1.6.3
attrs==25.3.0
autoflake==2.3.1
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
bson==0.5.10
bw-migrations==0.2
bw2calc==2.0.2
bw2data==4.4.4
bw2io==0.9.9
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constructive_geometries==1.0
country_converter==1.3
cssbeautifier==1.15.4
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
Deprecated==1.2.18
Django==5.2.1
django-extensions==4.1
django-unfold==0.59.0
djangorestframework==3.16.0
djlint==1.36.4
docopt==0.6.2
drf-yasg==1.21.10
EditorConfig==0.17.1
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fsspec==2025.5.1
greenlet==3.2.3
idna==3.10
ijson==3.4.0
inflection==0.5.1
isodate==0.7.2
isort==6.0.1
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
linear-tsv==1.1.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
matrix_utils==0.6
morefs==0.2.2
mrio-common-metadata==0.2.1
mypy_extensions==1.1.0
numba==0.61.2
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
peewee==3.18.1
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
prettytable==3.16.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyecospold==4.0.0
pyflakes==3.3.2
PyJWT==2.9.0
pyparsing==3.2.3
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.2
randonneur==0.6.1
randonneur_data==0.6
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==2.0.0
rpds-py==0.25.1
s3transfer==0.13.0
scipy==1.15.3
six==1.17.0
snowflake-id==1.0.2
SPARQLWrapper==2.0.0
sparse==0.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stats_arrays==0.7
structlog==25.4.0
tableschema==1.21.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unfold==1.2.3
unicodecsv==0.14.1
uritemplate==4.2.0
urllib3==2.4.0
voluptuous==0.15.2
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
wurst==0.4
xlrd==2.0.1
XlsxWriter==3.2.3
